Date Taken as DateTimeOffset (#8)
Properly parse date as a date. From DB into Model was easy, but from
Exiftool some trickery was needed. I initially tried to use the `parse`
utilities from the `time` crate, but it is tricky to implement a custom
format. I opted for regex as is more flexible, although more verbose.