Photos Cache (#16)
## Description

Caches Images to reduce AWS S3 fees.

1. Caches photos on request, if the image-size combination is not cached
it's stored in memory. If it's in the cache, then the request to AWS is
skipped.
2. Caches photos on bootstrap. This to avoid initial loading for the
main images.

## Known Issues

The current implementation has many problems, the ones I'm aware are:

1. Memory can grow infinitely: This could cause the droplet to run out
of memory, for now this is not an issue since I won't be caching hundred
of thousand of images. If this happens, then it should be addressed
later.
3. The photo is no longer streamed from AWS to the response: Since I'm a
Rust newbie, I could not make it work as I wanted, ideally I would be
able to stream the bytes from AWS into the browser and at the same time,
save the image into the cache as well. But the current approach first
downloads the full image and stores it in memory and in the cache, and
then is sent to the response. This does not bother me as much as it
should, since the images will be cached from the get-go.