Better random image (#23)
Instead of calculating a random image per request, which needs to query
the DB in every request. The random is instead calculated in the
frontend web-server, simply get all the `portfolio` photo ids, and then
randomize which `id` to get on every request.

This way, the amounts of queries gets decreased, which should result in
more requests per second without overloading the SQL connections.