Analytics (#39)
### Description

Implements a very basic analytics functionality. Completely anonymous
data is collected to have an idea of the number of requests and visits
made to the site. All of it is stored in memory and no personal
information is stored, not even IP Address or User Agent.

Each "unique" visitor is made with a UUID generated in the browser and
persisted with `localStorage` but nothing else is sent from the
front-end.