Cache images (#40)
Besides using the already cached images in memory, now the service
returns `304` when the client sends the `E-Tag` header that matches the
`md5` signature of the cached image. This avoids copying the `u8` bytes
around as well as saving network transfer, as the Browser skips
downloading the image altogether as it uses the browser cache.